requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
motor
pymupdf
crewai
crewai-tools
